+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Parameter to change label to given image.
 
+### Changed
+
+- Label and overview series can not be None.
+
 ## [0.6.0] - 2023-01-24
 
 ### Changed