+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Label and overview series can not be None.
+- Label and overview series can now be None.
 
 ## [0.6.0] - 2023-01-24