+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] -
 
+## [0.7.0] - 2023-02-13
+
 ### Added
 
 - Parameter to change label to given image.
@@ -110,7 +112,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release of wsidicom
 
-[Unreleased]: https://github.com/imi-bigpicture/wsidicom/compare/0.6.0..HEAD
+[Unreleased]: https://github.com/imi-bigpicture/wsidicom/compare/0.7.0..HEAD
+[0.7.0]: https://github.com/imi-bigpicture/wsidicom/compare/v0.6.0..v0.7.0
 [0.6.0]: https://github.com/imi-bigpicture/wsidicom/compare/v0.5.0..v0.6.0
 [0.5.0]: https://github.com/imi-bigpicture/wsidicom/compare/v0.4.0..v0.5.0
 [0.4.0]: https://github.com/imi-bigpicture/wsidicom/compare/v0.3.2..v0.4.0